Add token for existed api

+++ src/main/java/com/EntertainmentViet/backend/features/organizer/api/JobOfferController.java
@@ -4,10 +4,12 @@ import com.EntertainmentViet.backend.features.common.utils.RestUtils;
 import com.EntertainmentViet.backend.features.organizer.dto.JobOfferDto;
 import com.EntertainmentViet.backend.features.organizer.boundary.JobOfferBoundary;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.annotation.Async;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,6 +24,7 @@ import java.util.concurrent.CompletableFuture;
 @Validated
 @RequestMapping(path = JobOfferController.REQUEST_MAPPING_PATH)
 @RequiredArgsConstructor
+@Slf4j
 public class JobOfferController {
 
   public static final String REQUEST_MAPPING_PATH = "/organizers/{organizer_uid}/joboffers";
@@ -29,13 +32,26 @@ public class JobOfferController {
   private final JobOfferBoundary jobOfferService;
 
   @GetMapping()
-  public CompletableFuture<List<JobOfferDto>> findByOrganizerUid(@PathVariable("organizer_uid") UUID organizerUid) {
-    return CompletableFuture.completedFuture(jobOfferService.findByOrganizerUid(organizerUid));
+  public CompletableFuture<ResponseEntity<List<JobOfferDto>>> findByOrganizerUid(JwtAuthenticationToken token, @PathVariable("organizer_uid") UUID organizerUid) {
+
+    if (organizerUid != RestUtils.getUidFromToken(token)) {
+      log.warn(String.format("The token don't have enough access right to get information of organizer with uid '%s'", organizerUid));
+      return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.UNAUTHORIZED).build());
+    }
+
+    return CompletableFuture.completedFuture(ResponseEntity.ok().body(jobOfferService.findByOrganizerUid(organizerUid)));
   }
 
   @GetMapping(value = "/{uid}")
-  public CompletableFuture<ResponseEntity<JobOfferDto>> findByOrganizerUid(@PathVariable("organizer_uid") UUID organizerUid,
+  public CompletableFuture<ResponseEntity<JobOfferDto>> findByOrganizerUid(JwtAuthenticationToken token,
+                                                                           @PathVariable("organizer_uid") UUID organizerUid,
                                                                           @PathVariable("uid") UUID uid) {
+
+    if (organizerUid != RestUtils.getUidFromToken(token)) {
+      log.warn(String.format("The token don't have enough access right to get information of organizer with uid '%s'", organizerUid));
+      return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.UNAUTHORIZED).build());
+    }
+
     return CompletableFuture.completedFuture(jobOfferService.findByOrganizerUidAndUid(organizerUid, uid)
             .map( talentDto -> ResponseEntity
                     .ok()
@@ -47,9 +63,15 @@ public class JobOfferController {
   @PostMapping(
           consumes = MediaType.APPLICATION_JSON_VALUE,
           produces = MediaType.APPLICATION_JSON_VALUE)
-  @ResponseStatus(HttpStatus.CREATED)
-  public CompletableFuture<ResponseEntity<UUID>> create(HttpServletRequest request, @RequestBody @Valid JobOfferDto jobOfferDto,
+  public CompletableFuture<ResponseEntity<UUID>> create(JwtAuthenticationToken token, HttpServletRequest request,
+                                                        @RequestBody @Valid JobOfferDto jobOfferDto,
                                                         @PathVariable("organizer_uid") UUID organizerUid) {
+
+    if (organizerUid != RestUtils.getUidFromToken(token)) {
+      log.warn(String.format("The token don't have enough access right to get information of organizer with uid '%s'", organizerUid));
+      return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.UNAUTHORIZED).build());
+    }
+
     return  CompletableFuture.completedFuture(jobOfferService.create(jobOfferDto, organizerUid)
             .map(newJobOfferDto -> ResponseEntity
                     .created(RestUtils.getCreatedLocationUri(request, newJobOfferDto))
@@ -63,9 +85,14 @@ public class JobOfferController {
           consumes = MediaType.APPLICATION_JSON_VALUE,
           produces = MediaType.APPLICATION_JSON_VALUE,
           value = "/{uid}")
-  @ResponseStatus(HttpStatus.OK)
-  public CompletableFuture<ResponseEntity<UUID>> update(@RequestBody @Valid JobOfferDto jobOfferDto,
+  public CompletableFuture<ResponseEntity<UUID>> update(JwtAuthenticationToken token, @RequestBody @Valid JobOfferDto jobOfferDto,
                                                         @PathVariable("organizer_uid") UUID organizerUid, @PathVariable("uid") UUID uid) {
+
+    if (organizerUid != RestUtils.getUidFromToken(token)) {
+      log.warn(String.format("The token don't have enough access right to get information of organizer with uid '%s'", organizerUid));
+      return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.UNAUTHORIZED).build());
+    }
+
     return  CompletableFuture.completedFuture(jobOfferService.update(jobOfferDto, organizerUid, uid)
             .map(newJobOfferDto -> ResponseEntity
                     .ok()
@@ -76,7 +103,13 @@ public class JobOfferController {
   }
 
   @DeleteMapping(value = "/{uid}")
-  public CompletableFuture<ResponseEntity<Void>> delete(@PathVariable("organizer_uid") UUID organizerUid, @PathVariable("uid") UUID uid) {
+  public CompletableFuture<ResponseEntity<Void>> delete(JwtAuthenticationToken token, @PathVariable("organizer_uid") UUID organizerUid, @PathVariable("uid") UUID uid) {
+
+    if (organizerUid != RestUtils.getUidFromToken(token)) {
+      log.warn(String.format("The token don't have enough access right to get information of organizer with uid '%s'", organizerUid));
+      return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.UNAUTHORIZED).build());
+    }
+
     if (jobOfferService.delete(uid, organizerUid)) {
       return  CompletableFuture.completedFuture(ResponseEntity.ok().build());
     }
